refactor: lazy load map page content
